user service and query hook renaming
