+++ client/src/lib/hooks/query/user/useMutateLogin.ts
@@ -1,18 +1,11 @@
-import api from "@/lib/fetch/api";
+import { postLogin } from "@/lib/fetch/user-service";
 import { mk, qk } from "@/lib/query-keys";
 import { localUser } from "@/lib/user-storage";
 import type { Data } from "@/types/query.types";
 import type { User, UserLogin } from "@t/data/user.types";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-async function postLogin(user: UserLogin) {
-	return api.post<{ user: UserLogin }, Data<"user", User>>({
-		url: "/auth/login",
-		body: { user }
-	});
-}
-
-export default function useLoginMutation() {
+export default function useMutateLogin() {
 	const client = useQueryClient();
 
 	return useMutation<Data<"user", User>, unknown, UserLogin>({
