+++ client/src/lib/hooks/useAuthentication.ts
@@ -1,12 +1,12 @@
-import useLoginMutation from "@lib/hooks/query/user/useLoginMutation";
-import useLogoutMutation from "@lib/hooks/query/user/useLogoutMutation";
-import useMeQuery from "@lib/hooks/query/user/useMeQuery";
+import useMutateLogin from "@/lib/hooks/query/user/useMutateLogin";
+import useMutateLogout from "@/lib/hooks/query/user/useMutateLogout";
+import useQueryMe from "@/lib/hooks/query/user/useQueryMe";
 
 export default function useAuthentication() {
-	const { mutate: login } = useLoginMutation();
-	const { mutate: logout } = useLogoutMutation();
+	const { mutate: login } = useMutateLogin();
+	const { mutate: logout } = useMutateLogout();
 
-	const { data } = useMeQuery();
+	const { data } = useQueryMe();
 	const currentUser = data?.user;
 
 	const isLoggedIn = !!currentUser;
